Version 0.9.0

+++ student_information.py
@@ -5,7 +5,7 @@ import sqlite3
 class Ui_StudentInformation(object):
     def setupUi(self, StudentInformation, MainMenu):
         StudentInformation.setObjectName("StudentInformation")
-        StudentInformation.resize(683, 500)
+        StudentInformation.resize(670, 425)
         StudentInformation.setStyleSheet(
             ".QWidget{background-color: #CBB1A0;border-radius: 10px}")
         StudentInformation.setWindowFlags(
@@ -418,21 +418,13 @@ class Ui_StudentInformation(object):
         if (button_to_change.text() == "Add"):
             self.input_search.setEnabled(False)
             self.search_button.setEnabled(False)
-            self.input_student_id.setEnabled(True)
-            self.input_firstname.setEnabled(True)
-            self.input_lastname.setEnabled(True)
-            self.input_section.setEnabled(True)
-            self.input_cb_grade_level.setEnabled(True)
+            self.enable_inputs()
             self.clear_data_on_inputs()
             self.confirm_button.setText("Add Student")
         else:
             self.input_search.setEnabled(True)
             self.search_button.setEnabled(True)
-            self.input_student_id.setEnabled(False)
-            self.input_firstname.setEnabled(False)
-            self.input_lastname.setEnabled(False)
-            self.input_section.setEnabled(False)
-            self.input_cb_grade_level.setEnabled(False)
+            self.disable_inputs()
             self.clear_data_on_inputs()
             if (button_to_change.text() == "Edit"):
                 self.confirm_button.setText("Edit Student")
@@ -441,13 +433,13 @@ class Ui_StudentInformation(object):
 
     def clear_data_on_inputs(self):
         self.status_label.setText("")
-        self.input_student_id.setText("")
-        self.input_firstname.setText("")
-        self.input_lastname.setText("")
-        self.input_section.setText("")
-        self.input_search.setText("")
+        self.input_student_id.clear()
+        self.input_firstname.clear()
+        self.input_lastname.clear()
+        self.input_section.clear()
+        self.input_search.clear()
         self.input_cb_grade_level.setCurrentText("Grade 7")
-        self.status_label.setText("")
+        self.status_label.clear()
 
     def enable_inputs(self):
         self.input_student_id.setEnabled(True)
@@ -457,11 +449,11 @@ class Ui_StudentInformation(object):
         self.input_cb_grade_level.setEnabled(True)
 
     def disable_inputs(self):
-        self.input_student_id.setEnabled(False)
-        self.input_firstname.setEnabled(False)
-        self.input_lastname.setEnabled(False)
-        self.input_section.setEnabled(False)
-        self.input_cb_grade_level.setEnabled(False)
+        self.input_student_id.setDisabled(True)
+        self.input_firstname.setDisabled(True)
+        self.input_lastname.setDisabled(True)
+        self.input_section.setDisabled(True)
+        self.input_cb_grade_level.setDisabled(True)
 
     def search_student(self):
         search_value = self.input_search.text()
@@ -509,7 +501,8 @@ class Ui_StudentInformation(object):
             self.update_student_data(
                 student_id, firstname, lastname, section, grade_level)
         if (delete_button.isChecked()):
-            self.delete_student_data(student_id)
+            self.delete_student_data(
+                student_id, firstname, lastname, section, grade_level)
 
     def validate_data_to_add(self, student_id, firstname, lastname, section, grade_level):
         if (len(student_id) == 0 or len(firstname) == 0 or len(lastname) == 0 or len(section) == 0):
@@ -591,28 +584,40 @@ class Ui_StudentInformation(object):
         elif (result == QtWidgets.QMessageBox.StandardButton.No):
             pass
 
-    def delete_student_data(self, student_id):
-        # * Step 1: Initialize Database
-        con = sqlite3.connect('./db/test.db')
-        cur = con.cursor()
-
-        # * Step 2: Query the data to be deleted
-        query_delete = """
-        DELETE FROM STUDENT
-        WHERE Student_ID = ?;
-        """
-        interpolate_data = [student_id]
-        # * Step 3: Execute, Commit, Close
-        cur.execute(query_delete, interpolate_data)
-        con.commit()
-        con.close()
-        self.informative_message(
-            text="Data Deleted Successfully!",
-            subtext="You can still delete data after this message.",
-            window_title="Deleted Successfully",
-            icon_type="information"
-        )
-        self.clear_data_on_inputs()
+    def delete_student_data(self, student_id, firstname, lastname, section, grade_level):
+        msg = QtWidgets.QMessageBox()
+        msg.setIcon(QtWidgets.QMessageBox.Icon.Question)
+        msg.setText(
+            f"Are you sure to delete {student_id}? It will be gone forever...")
+        msg.setInformativeText(
+            f"Student ID:\t{student_id}\nFirstname:\t{firstname}\nLastname:\t{lastname}\nSection:\t\t{section}\nGrade Level:\t{grade_level}\n\nAre you sure?")
+        msg.setWindowTitle("DELETE Book Information Confirmation")
+        msg.setStandardButtons(
+            QtWidgets.QMessageBox.StandardButton.Yes | QtWidgets.QMessageBox.StandardButton.No)
+        result = msg.exec()
+        if (result == QtWidgets.QMessageBox.StandardButton.Yes):
+            # * Step 1: Initialize Database
+            con = sqlite3.connect('./db/test.db')
+            cur = con.cursor()
+            # * Step 2: Query the data to be deleted
+            query_delete = """
+            DELETE FROM STUDENT
+            WHERE Student_ID = ?;
+            """
+            interpolate_data = [student_id]
+            # * Step 3: Execute, Commit, Close
+            cur.execute(query_delete, interpolate_data)
+            con.commit()
+            con.close()
+            self.informative_message(
+                text="Data Deleted Successfully!",
+                subtext="You can still delete data after this message.",
+                window_title="Deleted Successfully",
+                icon_type="information"
+            )
+            self.clear_data_on_inputs()
+        elif (result == QtWidgets.QMessageBox.StandardButton.No):
+            pass
 
     def informative_message(self, text, subtext, window_title, icon_type="critical"):
         msg = QtWidgets.QMessageBox()
